reorganise test folder into unit (fast) and integration (slow) tests
