+++ Tests/unit/test_measurements.py
@@ -1,7 +1,9 @@
-from qtmWebGaitReport.measurements import Measurements
 from pathlib import Path
 
-working_directory = Path("TestFiles","ForMeasurementSectionTest","processed")
+from qtmWebGaitReport.measurements import Measurements
+
+working_directory = Path("TestFiles", "ForMeasurementSectionTest", "processed").absolute()
+
 
 def test_that_each_measurement_has_exactly_2_video_resources():
     meas = Measurements(str(working_directory))
@@ -12,4 +14,4 @@ def test_that_each_measurement_has_exactly_2_video_resources():
             if resource["type"] == "video":
                 video_resources.append(resource)
         assert len(video_resources) == 2, "num_video_resources {} != 2".format(len(video_resources))
-        
+
